Extract ParentImpl
